Create diameter of binary tree
Given the root of a binary tree, return the length of the diameter of the tree.

The diameter of a binary tree is the length of the longest path between any two nodes in a tree. This path may or may not pass through the root.

The length of a path between two nodes is represented by the number of edges between them.
Input: root = [1,2,3,4,5]
Output: 3
Explanation: 3 is the length of the path [4,2,1,3] or [5,2,1,3].

we do this by creating a stack and adding elements from root traversing in the left direction until it reaches leaf. After that we pop out the leaf and put it and its depth in a hashmap. We do the same for right subtrees. then we find the sum of leftdepth and rightdepth and it is our diameter.